requirements.txt
asgiref==3.2.7 
django==3.0.6 
pytz==2020.1 
sqlparse==0.3.1
certifi==2020.4.5.1 
chardet==3.0.4 
defusedxml==0.6.0 
django-allauth==0.41.0 
idna==2.9 
oauthlib==3.1.0 
python3-openid==3.1.0 
requests==2.23.0 
requests-oauthlib==1.3.0 
urllib3==1.25.9
psycopg2==2.8.5
django-multi-email-field==0.6.1
